ocr-service/requirements.txt
paddlepaddle==2.6.2
paddleocr==2.7.3
flask==3.0.0
flask-cors==4.0.0
pillow==10.2.0
numpy==1.26.4
